+++ apps/decodex/src/orchestrator/tests/intake_run_and_prompting/intake_run_prompting_dispatch/retained_lanes.rs
@@ -1,0 +1,275 @@
+use crate::{
+	orchestrator::{
+		self, IssueDispatchMode, TargetIssueRunContext,
+		tests::{self, FakeTracker, intake_run_and_prompting, recovery_terminal_support},
+	},
+	state::{self, StateStore},
+	worktree::WorktreeManager,
+};
+
+#[test]
+fn targeted_identifier_dispatch_accepts_status_visible_retained_closeout_lane() {
+	let (temp_dir, base_config, workflow) = tests::temp_project_layout();
+	let config = tests::service_config_with_github_token_env_var(&base_config, "HOME");
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = intake_run_and_prompting::run_and_prompting_service_owned_issue("In Review");
+	let tracker = FakeTracker::new(vec![issue.clone()]);
+	let worktree_manager =
+		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
+	let worktree =
+		worktree_manager.ensure_worktree(&issue.identifier, false).expect("worktree should exist");
+	let head_oid = tests::git_output(&worktree.path, &["rev-parse", "HEAD"]);
+	let pr_url = "https://github.com/hack-ink/decodex/pull/181";
+
+	state_store
+		.upsert_worktree(
+			config.service_id(),
+			&issue.id,
+			&worktree.branch_name,
+			&worktree.path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&worktree.path,
+		&tests::sample_review_handoff_marker(&worktree.branch_name, pr_url, &head_oid),
+	);
+
+	let _path_guard = recovery_terminal_support::install_fake_merged_pr_gh_response(
+		&temp_dir, &worktree, pr_url, &head_oid,
+	);
+	let snapshot = orchestrator::build_live_operator_status_snapshot(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		10,
+	)
+	.expect("status snapshot should build");
+	let lane = snapshot
+		.post_review_lanes
+		.iter()
+		.find(|lane| lane.issue_identifier == issue.identifier)
+		.expect("retained closeout lane should appear in status");
+
+	assert_eq!(lane.classification, "continue");
+	assert_eq!(lane.reason, "pull_request_merged_closeout_pending");
+	assert_eq!(lane.pr_state.as_deref(), Some("MERGED"));
+
+	let summary =
+		orchestrator::run_target_issue_once_with_inferred_dispatch(TargetIssueRunContext {
+			tracker: &tracker,
+			project: &config,
+			workflow: &workflow,
+			state_store: &state_store,
+			issue_id: &issue.identifier,
+			preferred_issue_state: None,
+			preferred_initial_issue_state: None,
+			dry_run: true,
+			lease_preacquired: false,
+			preferred_issue_claim_fd: None,
+			preferred_dispatch_slot_fd: None,
+			preferred_dispatch_slot_index: None,
+			dispatch_mode: IssueDispatchMode::Normal,
+			preferred_run_identity: None,
+			preferred_retry_budget_base: None,
+		})
+		.expect("targeted retained closeout identifier run should succeed")
+		.expect("status-visible retained closeout lane should dispatch by identifier");
+
+	assert_eq!(summary.issue_id, issue.id);
+	assert_eq!(summary.issue_identifier, issue.identifier);
+	assert_eq!(summary.dispatch_mode, IssueDispatchMode::Closeout);
+	assert_eq!(summary.run_id, "run-1");
+	assert_eq!(summary.attempt_number, 1);
+}
+
+#[test]
+fn targeted_identifier_dispatch_accepts_status_visible_review_repair_lane() {
+	let (temp_dir, base_config, workflow) = tests::temp_project_layout();
+	let config = tests::service_config_with_github_token_env_var(&base_config, "HOME");
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = intake_run_and_prompting::run_and_prompting_service_owned_issue("In Review");
+	let tracker = FakeTracker::new(vec![issue.clone()]);
+	let worktree_manager =
+		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
+	let worktree =
+		worktree_manager.ensure_worktree(&issue.identifier, false).expect("worktree should exist");
+	let head_oid = tests::git_output(&worktree.path, &["rev-parse", "HEAD"]);
+	let pr_url = "https://github.com/hack-ink/decodex/pull/184";
+
+	state_store
+		.upsert_worktree(
+			config.service_id(),
+			&issue.id,
+			&worktree.branch_name,
+			&worktree.path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&worktree.path,
+		&tests::sample_review_handoff_marker(&worktree.branch_name, pr_url, &head_oid),
+	);
+
+	let _path_guard = recovery_terminal_support::install_fake_conflicting_pr_gh_response(
+		&temp_dir, &worktree, pr_url, &head_oid,
+	);
+	let snapshot = orchestrator::build_live_operator_status_snapshot(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		10,
+	)
+	.expect("status snapshot should build");
+	let lane = snapshot
+		.post_review_lanes
+		.iter()
+		.find(|lane| lane.issue_identifier == issue.identifier)
+		.expect("retained repair lane should appear in status");
+
+	assert_eq!(lane.classification, "needs_review_repair");
+	assert_eq!(lane.reason, "pull_request_merge_conflict");
+
+	let summary =
+		orchestrator::run_target_issue_once_with_inferred_dispatch(TargetIssueRunContext {
+			tracker: &tracker,
+			project: &config,
+			workflow: &workflow,
+			state_store: &state_store,
+			issue_id: &issue.identifier,
+			preferred_issue_state: None,
+			preferred_initial_issue_state: None,
+			dry_run: true,
+			lease_preacquired: false,
+			preferred_issue_claim_fd: None,
+			preferred_dispatch_slot_fd: None,
+			preferred_dispatch_slot_index: None,
+			dispatch_mode: IssueDispatchMode::Normal,
+			preferred_run_identity: None,
+			preferred_retry_budget_base: None,
+		})
+		.expect("targeted retained repair identifier run should succeed")
+		.expect("status-visible retained repair lane should dispatch by identifier");
+
+	assert_eq!(summary.issue_id, issue.id);
+	assert_eq!(summary.issue_identifier, issue.identifier);
+	assert_eq!(summary.dispatch_mode, IssueDispatchMode::ReviewRepair);
+	assert_eq!(summary.issue_state, "In Review");
+}
+
+#[test]
+fn targeted_identifier_dispatch_accepts_stopped_active_closeout_lease() {
+	let (temp_dir, base_config, workflow) = tests::temp_project_layout();
+	let config = tests::service_config_with_github_token_env_var(&base_config, "HOME");
+	let state_store = StateStore::open_in_memory().expect("state store should open");
+	let issue = intake_run_and_prompting::run_and_prompting_service_owned_issue("Done");
+	let tracker = FakeTracker::new(vec![issue.clone()]);
+	let worktree_manager =
+		WorktreeManager::new(config.service_id(), config.repo_root(), config.worktree_root());
+	let worktree =
+		worktree_manager.ensure_worktree(&issue.identifier, false).expect("worktree should exist");
+	let head_oid = tests::git_output(&worktree.path, &["rev-parse", "HEAD"]);
+	let pr_url = "https://github.com/hack-ink/decodex/pull/183";
+
+	state_store
+		.record_run_attempt("run-1", &issue.id, 1, "running")
+		.expect("stopped running run attempt should record");
+	state_store
+		.upsert_lease(config.service_id(), &issue.id, "run-1", "Done")
+		.expect("stopped closeout lease should record");
+	state_store
+		.upsert_worktree(
+			config.service_id(),
+			&issue.id,
+			&worktree.branch_name,
+			&worktree.path.display().to_string(),
+		)
+		.expect("worktree should record");
+
+	state::write_run_activity_marker_for_process(&worktree.path, "run-1", 1, u32::MAX)
+		.expect("stopped closeout activity marker should write");
+	tests::seed_review_handoff_marker_for_path(
+		&state_store,
+		config.service_id(),
+		&worktree.path,
+		&tests::sample_review_handoff_marker(&worktree.branch_name, pr_url, &head_oid),
+	);
+
+	let _path_guard = recovery_terminal_support::install_fake_merged_pr_gh_response(
+		&temp_dir, &worktree, pr_url, &head_oid,
+	);
+	let snapshot = orchestrator::build_live_operator_status_snapshot(
+		&tracker,
+		&config,
+		&workflow,
+		&state_store,
+		10,
+	)
+	.expect("status snapshot should build");
+	let lane = snapshot
+		.post_review_lanes
+		.iter()
+		.find(|lane| lane.issue_identifier == issue.identifier)
+		.expect("retained closeout lane should appear in status");
+	let target_context = |dispatch_mode| {
+		intake_run_and_prompting::run_and_prompting_target_context(
+			&tracker,
+			&config,
+			&workflow,
+			&state_store,
+			&issue.identifier,
+			dispatch_mode,
+		)
+	};
+
+	assert_eq!(lane.classification, "continue");
+	assert_eq!(lane.reason, "pull_request_merged_closeout_pending");
+	assert_eq!(lane.pr_state.as_deref(), Some("MERGED"));
+	assert_eq!(snapshot.current_lanes.len(), 1);
+	assert_eq!(snapshot.current_lanes[0].process_alive, Some(false));
+	assert!(
+		orchestrator::issue_passes_closeout_dispatch_policy(
+			&tracker,
+			&issue,
+			&config,
+			&workflow,
+			&state_store,
+		)
+		.expect("closeout dispatch policy should evaluate"),
+		"stopped active closeout lease should still satisfy closeout policy"
+	);
+	assert!(
+		!orchestrator::target_issue_active_claim_blocks_dispatch(
+			&target_context(IssueDispatchMode::Closeout),
+			&issue.id,
+			&issue,
+		)
+		.expect("active closeout claim guard should evaluate"),
+		"stopped active closeout lease should not block closeout dispatch"
+	);
+
+	let explicit_summary =
+		orchestrator::run_target_issue_once(target_context(IssueDispatchMode::Closeout))
+			.expect("explicit retained closeout identifier run should succeed")
+			.expect("explicit closeout should accept the stopped active closeout lease");
+
+	assert_eq!(explicit_summary.dispatch_mode, IssueDispatchMode::Closeout);
+
+	let summary = orchestrator::run_target_issue_once_with_inferred_dispatch(target_context(
+		IssueDispatchMode::Normal,
+	))
+	.expect("targeted retained closeout identifier run should succeed")
+	.expect("stopped active closeout lease should not hide the closeout candidate");
+
+	assert_eq!(summary.issue_id, issue.id);
+	assert_eq!(summary.issue_identifier, issue.identifier);
+	assert_eq!(summary.dispatch_mode, IssueDispatchMode::Closeout);
+	assert_eq!(summary.run_id, "run-1");
+	assert_eq!(summary.attempt_number, 1);
+}
