+++ apps/decodex/src/orchestrator/tests/operator/status/text/program_readback.rs
@@ -1,294 +1,18 @@
+mod live_mapping;
+mod summary_fields;
+
 use crate::{
 	execution_program::{
-		ExecutionConflictDomain, ExecutionConflictDomainKind, ExecutionLinearIssueMapping,
-		ExecutionProgram, ExecutionProgramDependency, ExecutionProgramNode,
-		ExecutionProgramNodeStage, ExecutionQueueIntent,
+		ExecutionLinearIssueMapping, ExecutionProgram, ExecutionProgramDependency,
+		ExecutionProgramNode, ExecutionProgramNodeStage, ExecutionQueueIntent,
 	},
 	loop_contract::{DecisionPromotion, DecisionPromotionActorKind},
 	orchestrator::tests::operator::status::{
-		self, Connection, DecisionContract, FakeTracker, HashMap,
-		OperatorCodexAccountControlStatus, OperatorExecutionProgramNodeStatus,
-		OperatorExecutionProgramStatus, OperatorStatusSnapshot, ReviewHandoffMarker, ServiceConfig,
-		StateStore, Value, WorkflowDocument, env, orchestrator,
+		self, Connection, DecisionContract, FakeTracker, HashMap, OperatorExecutionProgramStatus,
+		OperatorStatusSnapshot, ServiceConfig, StateStore, Value, WorkflowDocument, env,
+		orchestrator,
 	},
 };
-
-#[test]
-fn operator_status_text_surfaces_execution_program_summary() {
-	let snapshot = OperatorStatusSnapshot {
-		project_id: String::from("decodex"),
-		run_limit: 10,
-		status_source: None,
-		snapshot_age_seconds: None,
-		warnings: Vec::new(),
-		warning_details: Vec::new(),
-		connector_backoffs: Vec::new(),
-		projects: Vec::new(),
-		account_control: OperatorCodexAccountControlStatus {
-			mode: String::from("balanced"),
-			account_selector: None,
-		},
-		accounts: Vec::new(),
-		current_lanes: Vec::new(),
-		queued_candidates: Vec::new(),
-		recent_runs: Vec::new(),
-		history_lanes: Vec::new(),
-		execution_programs: vec![OperatorExecutionProgramStatus {
-			program_id: String::from("program-853"),
-			status: String::from("blocked"),
-			source_contract_id: Some(String::from("contract-852")),
-			intake_kind: Some(String::from("goal_intake")),
-			public_summary: Some(String::from("Resolve promoted program work.")),
-			node_count: 3,
-			planned_count: 0,
-			mapped_count: 0,
-			ready_count: 1,
-			queued_count: 0,
-			blocked_count: 1,
-			held_count: 0,
-			active_count: 0,
-			needs_attention_count: 0,
-			completed_count: 1,
-			stale_count: 0,
-			superseded_count: 0,
-			dispatchable_count: 0,
-			mapped_issue_identifiers: vec![String::from("XY-853")],
-			node_readbacks: vec![OperatorExecutionProgramNodeStatus {
-				program_stage: String::from("runtime"),
-				lifecycle_state: String::from("blocked"),
-				readiness_state: String::from("blocked"),
-				issue_identifier: Some(String::from("XY-853")),
-				issue_state: Some(String::from("Todo")),
-				dispatch_action: None,
-				reason_codes: vec![String::from("dependency_not_terminal")],
-				reasons: vec![String::from(
-					"a dependency has not reached a required terminal state",
-				)],
-				next_action: String::from(
-					"Complete the dependency issue or refresh the Execution Program dependency plan if this remains stale.",
-				),
-			}],
-			readback_warning: None,
-		}],
-		worktrees: Vec::new(),
-		post_review_lanes: Vec::new(),
-	};
-	let rendered = orchestrator::render_operator_status(&snapshot);
-
-	assert!(rendered.contains("Execution programs: 1"));
-	assert!(rendered.contains("Execution Programs"));
-	assert!(rendered.contains(
-		"program_id: program-853 status=blocked source_contract_id: contract-852 intake_kind=goal_intake summary=\"Resolve promoted program work.\" nodes=3 planned=0 mapped=0 ready=1 queued=0 blocked=1 held=0 active=0 attention=0 completed=1 stale=0 superseded=0 dispatchable=0 mapped_issues=XY-853"
-	));
-	assert!(rendered.contains(
-		"node: issue=XY-853 issue_state=Todo program_stage=runtime lifecycle=blocked readiness=blocked dispatch_action=none reason_codes=dependency_not_terminal reasons=\"a dependency has not reached a required terminal state\" next_action=\"Complete the dependency issue or refresh the Execution Program dependency plan if this remains stale.\""
-	));
-}
-
-#[test]
-fn operator_status_json_uses_direct_dispatch_program_fields() {
-	let snapshot: OperatorStatusSnapshot = serde_json::from_value(serde_json::json!({
-		"project_id": "decodex",
-		"run_limit": 10,
-		"warnings": [],
-		"warning_details": [],
-		"connector_backoffs": [],
-		"projects": [],
-		"account_control": {
-			"mode": "balanced",
-			"account_selector": null,
-		},
-		"accounts": [],
-		"current_lanes": [],
-		"recent_runs": [],
-		"history_lanes": [],
-		"execution_programs": [{
-			"program_id": "direct-dispatch-program",
-			"source_contract_id": null,
-			"node_count": 1,
-			"planned_count": 0,
-			"mapped_count": 0,
-			"ready_count": 1,
-			"queued_count": 0,
-			"blocked_count": 0,
-			"held_count": 0,
-			"active_count": 0,
-			"needs_attention_count": 0,
-			"completed_count": 0,
-			"stale_count": 0,
-			"superseded_count": 0,
-			"dispatchable_count": 1,
-			"mapped_issue_identifiers": ["XY-853"],
-			"node_readbacks": [{
-				"lifecycle_state": "ready",
-				"readiness_state": "ready",
-				"issue_identifier": "XY-853",
-				"issue_state": "Todo",
-				"dispatch_action": "dispatch",
-				"reason_codes": ["ready_for_linear_execution"],
-				"reasons": ["node is ready for normal Linear issue execution"],
-				"next_action": "The program scheduler can dispatch this node directly."
-			}],
-			"readback_warning": null,
-		}],
-		"queued_candidates": [],
-		"worktrees": [],
-		"post_review_lanes": [],
-	}))
-	.expect("operator snapshot should deserialize");
-	let program = snapshot.execution_programs.first().expect("program should deserialize");
-
-	assert_eq!(program.status, "unknown");
-	assert_eq!(program.intake_kind, None);
-	assert_eq!(program.public_summary, None);
-	assert_eq!(program.dispatchable_count, 1);
-	assert_eq!(program.node_readbacks[0].dispatch_action.as_deref(), Some("dispatch"));
-}
-
-#[test]
-fn operator_status_snapshot_surfaces_program_intake_and_node_readbacks() {
-	let (_temp_dir, config, workflow) = status::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-
-	seed_program_readback_status(&state_store, &config);
-
-	let snapshot = build_program_readback_snapshot(&config, &workflow, &state_store);
-	let program = snapshot.execution_programs.first().expect("program should surface");
-	let program_json = program_readback_json(&snapshot);
-
-	assert_program_readback_summary(program);
-	assert_program_readback_json(&program_json);
-	assert_program_node_readbacks(program, &program_json);
-}
-
-#[test]
-fn operator_status_program_readback_prefers_post_review_owner_over_stale_active_label() {
-	let (_temp_dir, config, workflow) = status::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let issue_id = "issue-post-review";
-	let issue_identifier = "PUB-946";
-	let branch_name = "x/pubfi-pub-946";
-	let conflict = ExecutionConflictDomain::new(ExecutionConflictDomainKind::Module, "runtime")
-		.expect("conflict domain should build");
-	let node =
-		status_program_active_node("node-post-review", issue_id, issue_identifier, "In Review")
-			.with_conflict_domains([conflict])
-			.expect("conflict domain should attach");
-	let program = ExecutionProgram::from_issue_batch_intake(
-		"program-post-review-owner",
-		config.service_id(),
-		"program-post-review-owner-fingerprint",
-		"Track post-review owner readback.",
-		vec![node],
-	)
-	.expect("program should build");
-
-	state_store
-		.upsert_execution_program(config.service_id(), program)
-		.expect("program should persist");
-	state_store
-		.upsert_review_handoff_marker(
-			config.service_id(),
-			issue_id,
-			&ReviewHandoffMarker::new(
-				"pub-946-attempt-1",
-				1,
-				branch_name,
-				"https://github.com/hack-ink/pubfi/pull/946",
-				"main",
-				branch_name,
-				"1111111111111111111111111111111111111111",
-			),
-		)
-		.expect("review lifecycle should persist");
-
-	let snapshot = build_program_readback_snapshot(&config, &workflow, &state_store);
-	let program = snapshot.execution_programs.first().expect("program should surface");
-	let node = program.node_readbacks.first().expect("post-review node should surface");
-
-	assert_eq!(program.status, "active");
-	assert_eq!(program.active_count, 1);
-	assert_eq!(program.blocked_count, 0);
-	assert_eq!(node.lifecycle_state, "post_review");
-	assert_eq!(node.readiness_state, "blocked");
-	assert!(node.reason_codes.contains(&String::from("mapped_issue_post_review_owner")));
-	assert!(!node.reason_codes.contains(&String::from("mapped_issue_active_label_present")));
-	assert!(!node.reason_codes.contains(&String::from("conflict_domain_occupied")));
-	assert_eq!(
-		node.reasons,
-		vec![String::from(
-			"Review & Landing owns this issue until post-review landing or closeout finishes",
-		)]
-	);
-	assert_eq!(
-		node.next_action,
-		"Continue the retained post-review lifecycle before dispatching this program node."
-	);
-}
-
-#[test]
-fn operator_status_program_readback_refreshes_live_tracker_issue_mapping() {
-	let (_temp_dir, config, workflow) = status::temp_project_layout();
-	let state_store = StateStore::open_in_memory().expect("state store should open");
-	let stale_mapping = status_program_issue_mapping("issue-live-refresh", "PUB-1597", "Todo")
-		.with_needs_attention_label(true);
-	let node = ExecutionProgramNode::new(
-		"node-live-refresh",
-		ExecutionProgramNodeStage::Runtime,
-		"Close stale Program attention after the mapped issue is terminal.",
-		ExecutionQueueIntent::ReadyToQueue,
-	)
-	.expect("node should build")
-	.with_acceptance_expectations(["The mapped issue reflects live tracker state."])
-	.expect("acceptance should attach")
-	.with_validation_expectations(["Build operator status."])
-	.expect("validation should attach")
-	.with_linear_issue(stale_mapping)
-	.expect("stale mapping should attach");
-	let program = ExecutionProgram::from_issue_batch_intake(
-		"program-live-refresh",
-		config.service_id(),
-		"program-live-refresh-fingerprint",
-		"Refresh live Program issue metadata.",
-		vec![node],
-	)
-	.expect("program should build");
-
-	state_store
-		.upsert_execution_program(config.service_id(), program)
-		.expect("program should persist");
-
-	let live_issue = status::sample_issue_with_sort_fields(
-		"issue-live-refresh",
-		"PUB-1597",
-		"Done",
-		&[],
-		Some(1),
-		"2026-06-19T00:00:00.000Z",
-	);
-	let tracker = FakeTracker::new(vec![live_issue]);
-	let snapshot = orchestrator::build_live_operator_status_snapshot(
-		&tracker,
-		&config,
-		&workflow,
-		&state_store,
-		10,
-	)
-	.expect("status snapshot should build");
-	let program = snapshot.execution_programs.first().expect("program should surface");
-
-	assert_eq!(program.status, "completed");
-	assert_eq!(program.completed_count, 1);
-	assert_eq!(program.needs_attention_count, 0);
-	assert_eq!(program.blocked_count, 0);
-	assert_eq!(program.dispatchable_count, 0);
-	assert!(
-		program.node_readbacks.is_empty(),
-		"terminal refreshed Program nodes should not render stale attention readbacks"
-	);
-	assert_eq!(tracker.refresh_queries.borrow().len(), 1);
-	assert_eq!(tracker.refresh_queries.borrow()[0], vec![String::from("issue-live-refresh")]);
-}
 
 fn seed_program_readback_status(state_store: &StateStore, config: &ServiceConfig) {
 	let program = ExecutionProgram::from_issue_batch_intake(
